Emphasize that WCMP2 is GeoJSON #190 (#193)
* Emphasize that WCMP2 is GeoJSON #190

For now, only examples are updated and one humble sentence added to the normative text (chapter 7.1.1. Overview).
I was not able to find a place where we would talk about the WCMP record "representation". Should I add something like "7.1.2 WCMP record representation" (which would shift all other things in 7.1) with a sentence or two and a requirement?

* Add WCMP media type as REC_media_type.adoc #190

---------

Co-authored-by: Tom Kralidis <tomkralidis@gmail.com>